+++ src/server.rs
@@ -1,4 +1,5 @@
 use crate::LambdaRuntimeApiClient;
+use http_body_util::{BodyExt, Full};
 use hyper::{
   body::{Body, Incoming},
   server::conn::http1,
@@ -67,7 +68,22 @@ impl MockLambdaRuntimeApiServer {
   /// forwarding requests to a new [`LambdaRuntimeApiClient`], and responding with the client's response.
   pub async fn passthrough(&self) {
     self
-      .serve(|req| async { LambdaRuntimeApiClient::new().await.send_request(req).await })
+      .serve(|req| async {
+        // tested and it looks like we create the client every time is faster than lock a Arc<Mutex<>> and reuse it.
+        // create a new client and send the request usually cost < 1ms.
+        let res = LambdaRuntimeApiClient::new()
+          .await
+          .send_request(req)
+          .await
+          .unwrap();
+        let (parts, body) = res.into_parts();
+        let bytes = body.collect().await.unwrap().to_bytes();
+        Ok(Response::from_parts(parts, Full::new(bytes)))
+
+        // TODO: why we can't just return `LambdaRuntimeApiClient::new().await.send_request(req).await`?
+        // tested and it works but will add ~40ms latency when serving API GW event (maybe for all big event), why?
+        // maybe because of the `Incoming` type? can we stream the body instead of buffering it?
+      })
       .await
   }
 }