+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## v0.2.1
+
+- Feat: add `LambdaRuntimeApiClient::forward`.
+
 ## v0.2.0
 
 - **_Breaking Change_**: stricter trait bound for `MockLambdaRuntimeApiServer` methods.